# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ModelEntry": {
        "additionalProperties": true,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modality": {
            "enum": [
              "image",
              "video",
              "music"
            ],
            "title": "Modality",
            "type": "string"
          },
          "object": {
            "const": "model",
            "default": "model",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "id",
          "modality"
        ],
        "title": "ModelEntry",
        "type": "object"
      },
...
      "ModelListResponse": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "title": "Data",
            "type": "array"
          },
          "object": {
            "const": "list",
            "default": "list",
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "ModelListResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": true,
        "description": "RFC 9457 problem details with stable gateway extensions.",
        "properties": {
          "code": {
            "description": "Stable machine-readable gateway error code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable detail for this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Field-level validation errors, when applicable.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI reference identifying this occurrence.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request correlation identifier.",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP response status code.",
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "API key",
        "description": "Front-end API key sent as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified image, video, and music gateway with separate REST APIs.",
    "title": "mm-gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "parameters": [
          {
            "description": "Filter models by output modality.",
            "in": "query",
            "name": "modality",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "image",
                    "video",
                    "music"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter models by output modality.",
              "title": "Modality"
            }
          },
          {
            "description": "Bearer token: \"Bearer <api-key>\".",
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer token: \"Bearer <api-key>\".",
              "examples": [
                "Bearer sk-gateway-demo"
              ],
              "title": "Authorization"
            }
          },
          {
            "description": "Client-supplied request id (echoed back).",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client-supplied request id (echoed back).",
              "examples": [
                "req-01HZX4J3K7NQ8X2V9Y6R5W4T3P"
              ],
              "title": "X-Request-Id"
            }
          },
          {
            "description": "Previously returned ETag; unchanged resources return 304.",
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Previously returned ETag; unchanged resources return 304.",
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "gateway-image-pro",
                      "modality": "image",
                      "object": "model"
                    },
                    {
                      "id": "gateway-video-pro",
                      "modality": "video",
                      "object": "model"
                    },
                    {
                      "id": "gateway-music-lyria",
                      "modality": "music",
                      "object": "model"
                    }
                  ],
                  "object": "list"
                },
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "Models available to the authenticated client.",
            "headers": {
              "ETag": {
                "description": "Version identifier for conditional retrieval.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "The model catalogue has not changed."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to use any generation service"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Model or task not found."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/models",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Models",
        "tags": [
          "meta"
        ]
      }
    },
...
  }
}